refactor(proc): remove ProcReference
Keep this simpler; move the methods to InputReference, and remove this
from the processor.

Close: #21
