+++ bibliography/Cargo.toml
@@ -9,8 +9,8 @@ test = true
 doctest = true
 bench = true
 doc = true
-edition = "2021"       # The edition of the target.
-crate-type = ["lib"]   # The crate types to generate.
+edition = "2021" 
+crate-type = ["lib"] 
 
 [dependencies]
 schemars = "0.8.12"
@@ -18,3 +18,6 @@ serde = "1.0.162"
 serde_derive = "1.0.162"
 serde_json = "1.0.96"
 serde_yaml = "0.9.21"
+chrono = { version = "0.4", features = ["unstable-locales"] }
+style = { path = "../style", package = "csln-style" }
+
