+++ processor/Cargo.toml
@@ -19,7 +19,6 @@ serde_derive = "1.0.162"
 serde_json = "1.0.96"
 serde_yaml = "0.9.21"
 edtf = "0.2.0"
-chrono = { version = "0.4", features = ["unstable-locales"] }
 style = { path = "../style", package = "csln-style" }
 bibliography = { path = "../bibliography", package = "csln-bibliography" }
 itertools = "0.10.5"
