+++ processor/src/lib.rs
@@ -1,4 +1,3 @@
-use chrono::NaiveDate;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -9,6 +8,7 @@ use std::option::Option;
 use bibliography::InputBibliography as Bibliography;
 use bibliography::InputReference;
 use style::options::{SortOrder, StyleSortGroupKey, StyleSorting};
+#[allow(unused_imports)] // for now
 use style::template::{
     Contributors, DateForm, Dates, StyleTemplateComponent, StyleTemplateContributor,
 };
@@ -264,90 +264,5 @@ impl Processor {
             bibliography,
             locale,
         }
-    }
-}
-
-// REVIEW with recent changes, do we need this, or put the methods directly on InputReference?
-impl ProcReference {
-    fn format_names(names: Vec<String>) -> String {
-        let mut name_string = String::new();
-        if names.len() == 1 {
-            name_string = names[0].to_string();
-        } else if names.len() == 2 {
-            name_string = names.join(" and ");
-        } else if names.len() > 2 {
-            let last_author = names.last().unwrap();
-            let other_authors = &names[..names.len() - 1];
-            name_string = other_authors.join(", ");
-            name_string.push_str(", and ");
-            name_string.push_str(last_author);
-        }
-        name_string
-    }
-
-    pub fn format_contributors(&self, template_component: StyleTemplateContributor) -> String {
-        match template_component.contributor {
-            Contributors::Author => {
-                let authors = self
-                    .data
-                    .author
-                    .as_ref()
-                    .unwrap_or(&Vec::new())
-                    .iter()
-                    .map(|name| name.to_string())
-                    .collect::<Vec<String>>();
-                ProcReference::format_names(authors)
-            }
-            Contributors::Editor => {
-                let editors = self
-                    .data
-                    .editor
-                    .as_ref()
-                    .unwrap()
-                    .iter()
-                    .map(|editor| editor.to_string())
-                    .collect::<Vec<String>>();
-                ProcReference::format_names(editors)
-            }
-            Contributors::Translator => {
-                let translators = self
-                    .data
-                    .translator
-                    .as_ref()
-                    .unwrap()
-                    .iter()
-                    .map(|translator| translator.to_string())
-                    .collect::<Vec<String>>();
-                ProcReference::format_names(translators)
-            }
-            Contributors::Director => todo!(),
-            Contributors::Recipient => todo!(),
-            Contributors::Interviewer => todo!(),
-            Contributors::Interviewee => todo!(),
-            Contributors::Inventor => todo!(),
-            Contributors::Counsel => todo!(),
-            Contributors::Composer => todo!(),
-            Contributors::Publisher => todo!(),
-        }
-    }
-
-    pub fn format_date(&self, date: Dates, form: DateForm) -> String {
-        let date_string: &str = match date {
-            Dates::Issued => self.data.issued.as_ref().unwrap(),
-            Dates::Accessed => todo!(),
-            Dates::OriginalPublished => todo!(),
-        };
-
-        let format_string: &str = match form {
-            DateForm::Year => "%Y",
-            DateForm::YearMonth => "%Y-%m",
-            DateForm::Full => "%Y-%m-%d",
-            DateForm::MonthDay => "%m-%d",
-        };
-
-        // use EDTF instead?
-        let date: NaiveDate = NaiveDate::parse_from_str(date_string, "%Y-%m-%d").unwrap();
-        let formatted_date: String = date.format(format_string).to_string();
-        formatted_date
     }
 }
